pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "au-grants-agent"
version = "0.1.0"
description = "Autonomous AI agent that crawls Australian Government grants and generates PhD-level proposals"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "NullShift", email = "dev@nullshift.sh"},
]
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "anthropic>=0.40",
    "openai>=1.12",
    "rich>=13.7",
    "click>=8.1",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "python-docx>=1.1",
    "fpdf2>=2.7",
    "markdownify>=0.13",
]

[project.scripts]
au-grants = "au_grants_agent.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
